major refactor to add utils and jwt
