+++ src/config/passport/passport.config.js
@@ -1,0 +1,7 @@
+import passport from "passport";
+
+// Load strategies
+import "./local.strategy.js";
+import "./jwt.strategy.js";
+
+export default passport;
